spec: change weldr requirement
The "Suggests" hint does not always resolve our dependencies as desired
so, for new releases, cockpit-composer specifically requires
osbuild-composer. For distros that cockpit-composer has already released
into, the "weldr" group remains the requirement so that users already
running osbuild-composer as the backend can continue to use it instead
of lorax-composer. But, if osbuild-composer is unavailable, lorax is
installed instead.

Cherry-picked from master commit 6e020fd36